Move TextArea to flat structure
