+++ packages/client/src/components/form/FormFieldGenerator.tsx
@@ -13,10 +13,10 @@ import * as React from 'react'
 import {
   CheckboxGroup,
   RadioGroup,
-  TextArea,
   TextInput,
   RadioSize
 } from '@opencrvs/components/lib/forms'
+import { TextArea } from '@opencrvs/components/lib/TextArea'
 import { Select } from '@opencrvs/components/lib/Select'
 import { DateField } from '@opencrvs/components/lib/DateField'
 import { WarningMessage } from '@opencrvs/components/lib/WarningMessage'
