+++ packages/client/src/views/SysAdmin/Config/Forms/Wizard/SaveActionModal.tsx
@@ -32,7 +32,8 @@ import {
 import { buttonMessages } from '@client/i18n/messages'
 import { ResponsiveModal } from '@opencrvs/components/lib/interface'
 import { messages } from '@client/i18n/messages/views/formConfig'
-import { InputField, TextArea } from '@opencrvs/components/lib/forms'
+import { InputField } from '@opencrvs/components/lib/forms'
+import { TextArea } from '@opencrvs/components/lib/TextArea'
 import { useParams } from 'react-router'
 import { goToFormConfigHome } from '@client/navigation'
 import { updateFormConfig } from '@client/forms/configuration/formConfig/actions'
